Feat/auto turnoff display add ble power settings

+++ components/power_management/include/power_management.h
@@ -100,9 +100,9 @@ bool power_management_is_wakeup_by_touch(void);
 bool power_management_is_wakeup_from_timer(void);
 
 /**
- * @brief 主动进入 deep sleep（60 秒后定时唤醒 + 触摸中断唤醒）
+ * @brief 主动进入 deep sleep（指定毫秒后定时唤醒 + 触摸中断唤醒）
  */
-void power_management_enter_deepsleep(uint16_t wakeup_time_ms);
+void power_management_enter_deepsleep(uint32_t wakeup_time_ms);
 
 #ifdef __cplusplus
 }
